part 1 complete

+++ src/main/java/training/supportbank/Transaction.java
@@ -1,12 +1,14 @@
 package training.supportbank;
 
+import java.math.BigDecimal;
+
 public class Transaction {
 
     private final Account sender;
     private final Account receiver;
-    private final Float amount;
+    private final BigDecimal amount;
 
-    public Transaction (Account initSending, Account initReceiving, Float initAmmount) {
+    public Transaction (Account initSending, Account initReceiving, BigDecimal initAmmount) {
 
         sender = initSending;
         receiver = initReceiving;
@@ -14,7 +16,7 @@ public class Transaction {
 
     }
 
-    public float GetAmount() {
+    public BigDecimal GetAmount() {
         return amount;
     }
